+++ nest-forms-backend/src/form-sender/form-sender.module.ts
@@ -1,14 +1,21 @@
 import { Module } from '@nestjs/common'
 
+import ApiJwtTokensModule from '../api-jwt-tokens/api-jwt-tokens.module'
 import { AuthV2Module } from '../auth-v2/auth-v2.module'
+import BaConfigModule from '../config/ba-config.module'
+import ConvertPdfModule from '../convert-pdf/convert-pdf.module'
+import FilesModule from '../files/files.module'
+import FormValidatorRegistryModule from '../form-validator-registry/form-validator-registry.module'
 import FormsModule from '../forms/forms.module'
 import { FormsV2Module } from '../forms-v2/forms-v2.module'
 import NasesModule from '../nases/nases.module'
 import FormRegistrationStatusRepository from '../nases/utils-services/form-registration-status.repository'
 import PrismaModule from '../prisma/prisma.module'
+import RabbitmqClientModule from '../rabbitmq-client/rabbitmq-client.module'
 import ThrowerErrorGuard from '../utils/guards/thrower-error.guard'
 import { LineLoggerSubservice } from '../utils/subservices/line-logger.subservice'
 import FormSenderController from './form-sender.controller'
+import { FormSenderService } from './form-sender.service'
 
 @Module({
   imports: [
@@ -16,12 +23,19 @@ import FormSenderController from './form-sender.controller'
     FormsModule,
     FormsV2Module,
     AuthV2Module,
+    ApiJwtTokensModule,
+    BaConfigModule,
+    FilesModule,
+    ConvertPdfModule,
+    FormValidatorRegistryModule,
+    RabbitmqClientModule,
     PrismaModule,
   ],
   providers: [
     ThrowerErrorGuard,
     LineLoggerSubservice,
     FormRegistrationStatusRepository,
+    FormSenderService,
   ],
   controllers: [FormSenderController],
 })
