+++ nest-forms-backend/src/nases/nases.module.ts
@@ -1,24 +1,13 @@
 import { Module } from '@nestjs/common'
 
 import ApiJwtTokensModule from '../api-jwt-tokens/api-jwt-tokens.module'
-import UserInfoPipeModule from '../auth/decorators/user-info-pipe.module'
-import { AuthV2Module } from '../auth-v2/auth-v2.module'
 import ClientsModule from '../clients/clients.module'
 import BaConfigModule from '../config/ba-config.module'
 import ConvertModule from '../convert/convert.module'
-import ConvertPdfModule from '../convert-pdf/convert-pdf.module'
-import FilesModule from '../files/files.module'
-import FormValidatorRegistryModule from '../form-validator-registry/form-validator-registry.module'
-import FormsModule from '../forms/forms.module'
-import { FormsV2Module } from '../forms-v2/forms-v2.module'
 import PrismaModule from '../prisma/prisma.module'
-import RabbitmqClientModule from '../rabbitmq-client/rabbitmq-client.module'
-import ScannerClientService from '../scanner-client/scanner-client.service'
 import TaxModule from '../tax/tax.module'
 import ThrowerErrorGuard from '../utils/guards/thrower-error.guard'
-import { LineLoggerSubservice } from '../utils/subservices/line-logger.subservice'
 import MinioClientSubservice from '../utils/subservices/minio-client.subservice'
-import NasesController from './nases.controller'
 import NasesService from './nases.service'
 import NasesContactsService from './services/nases.contacts.service'
 import NasesSenderService from './services/nases.sender.service'
@@ -27,30 +16,20 @@ import NasesCronSubservice from './utils-services/nases.cron.subservice'
 
 @Module({
   imports: [
-    ApiJwtTokensModule,
-    PrismaModule,
-    FormsModule,
-    RabbitmqClientModule,
-    FilesModule,
-    ConvertModule,
-    TaxModule,
-    ConvertPdfModule,
-    FormValidatorRegistryModule,
     ClientsModule,
-    UserInfoPipeModule,
-    FormsV2Module,
-    AuthV2Module,
+    ApiJwtTokensModule,
     BaConfigModule,
+    ConvertModule,
+    PrismaModule,
+    TaxModule,
   ],
   providers: [
+    FormRegistrationStatusRepository,
+    MinioClientSubservice,
     NasesService,
     NasesSenderService,
     ThrowerErrorGuard,
-    ScannerClientService,
-    MinioClientSubservice,
-    LineLoggerSubservice,
     NasesCronSubservice,
-    FormRegistrationStatusRepository,
     NasesContactsService,
   ],
   exports: [
@@ -59,6 +38,5 @@ import NasesCronSubservice from './utils-services/nases.cron.subservice'
     NasesCronSubservice,
     NasesContactsService,
   ],
-  controllers: [NasesController],
 })
 export default class NasesModule {}
