+++ openapi-clients/forms/api.ts
@@ -3909,820 +3909,6 @@ export class HealthcheckApi extends BaseAPI {
 }
 
 /**
- * NasesApi - axios parameter creator
- */
-export const NasesApiAxiosParamCreator = function (configuration?: Configuration) {
-  return {
-    /**
-     * Archive form (hide from user but keep in database)
-     * @summary
-     * @param {string} formId
-     * @param {*} [options] Override http request option.
-     * @deprecated
-     * @throws {RequiredError}
-     */
-    nasesControllerDeleteForm: async (
-      formId: string,
-      options: RawAxiosRequestConfig = {},
-    ): Promise<RequestArgs> => {
-      // verify required parameter 'formId' is not null or undefined
-      assertParamExists('nasesControllerDeleteForm', 'formId', formId)
-      const localVarPath = `/nases/{formId}`.replace('{formId}', encodeURIComponent(String(formId)))
-      // use dummy base URL string because the URL constructor only accepts absolute URLs.
-      const localVarUrlObj = new URL(localVarPath, DUMMY_BASE_URL)
-      let baseOptions
-      if (configuration) {
-        baseOptions = configuration.baseOptions
-      }
-
-      const localVarRequestOptions = { method: 'DELETE', ...baseOptions, ...options }
-      const localVarHeaderParameter = {} as any
-      const localVarQueryParameter = {} as any
-
-      // authentication cognitoGuestIdentityId required
-      await setApiKeyToObject(localVarHeaderParameter, 'X-Cognito-Guest-Identity-Id', configuration)
-
-      // authentication bearer required
-      // http bearer authentication required
-      await setBearerAuthToObject(localVarHeaderParameter, configuration)
-
-      setSearchParams(localVarUrlObj, localVarQueryParameter)
-      let headersFromBaseOptions = baseOptions && baseOptions.headers ? baseOptions.headers : {}
-      localVarRequestOptions.headers = {
-        ...localVarHeaderParameter,
-        ...headersFromBaseOptions,
-        ...options.headers,
-      }
-
-      return {
-        url: toPathString(localVarUrlObj),
-        options: localVarRequestOptions,
-      }
-    },
-    /**
-     * Return form by ID and by logged user
-     * @summary
-     * @param {string} formId
-     * @param {*} [options] Override http request option.
-     * @deprecated
-     * @throws {RequiredError}
-     */
-    nasesControllerGetForm: async (
-      formId: string,
-      options: RawAxiosRequestConfig = {},
-    ): Promise<RequestArgs> => {
-      // verify required parameter 'formId' is not null or undefined
-      assertParamExists('nasesControllerGetForm', 'formId', formId)
-      const localVarPath = `/nases/form/{formId}`.replace(
-        '{formId}',
-        encodeURIComponent(String(formId)),
-      )
-      // use dummy base URL string because the URL constructor only accepts absolute URLs.
-      const localVarUrlObj = new URL(localVarPath, DUMMY_BASE_URL)
-      let baseOptions
-      if (configuration) {
-        baseOptions = configuration.baseOptions
-      }
-
-      const localVarRequestOptions = { method: 'GET', ...baseOptions, ...options }
-      const localVarHeaderParameter = {} as any
-      const localVarQueryParameter = {} as any
-
-      // authentication cognitoGuestIdentityId required
-      await setApiKeyToObject(localVarHeaderParameter, 'X-Cognito-Guest-Identity-Id', configuration)
-
-      // authentication bearer required
-      // http bearer authentication required
-      await setBearerAuthToObject(localVarHeaderParameter, configuration)
-
-      localVarHeaderParameter['Accept'] = 'application/json'
-
-      setSearchParams(localVarUrlObj, localVarQueryParameter)
-      let headersFromBaseOptions = baseOptions && baseOptions.headers ? baseOptions.headers : {}
-      localVarRequestOptions.headers = {
-        ...localVarHeaderParameter,
-        ...headersFromBaseOptions,
-        ...options.headers,
-      }
-
-      return {
-        url: toPathString(localVarUrlObj),
-        options: localVarRequestOptions,
-      }
-    },
-    /**
-     * Get paginated forms
-     * @summary Get paginated forms
-     * @param {string} [currentPage] Page number
-     * @param {string} [pagination] Number of items per page
-     * @param {Array<FormState>} [states] Forms in which states are searched - when omitted, all forms of the user are searched
-     * @param {boolean} [userCanEdit] Get only forms in such a state, that user can still edit it.
-     * @param {string} [formDefinitionSlug] Slug of the form definition
-     * @param {*} [options] Override http request option.
-     * @deprecated
-     * @throws {RequiredError}
-     */
-    nasesControllerGetForms: async (
-      currentPage?: string,
-      pagination?: string,
-      states?: Array<FormState>,
-      userCanEdit?: boolean,
-      formDefinitionSlug?: string,
-      options: RawAxiosRequestConfig = {},
-    ): Promise<RequestArgs> => {
-      const localVarPath = `/nases/forms`
-      // use dummy base URL string because the URL constructor only accepts absolute URLs.
-      const localVarUrlObj = new URL(localVarPath, DUMMY_BASE_URL)
-      let baseOptions
-      if (configuration) {
-        baseOptions = configuration.baseOptions
-      }
-
-      const localVarRequestOptions = { method: 'GET', ...baseOptions, ...options }
-      const localVarHeaderParameter = {} as any
-      const localVarQueryParameter = {} as any
-
-      // authentication bearer required
-      // http bearer authentication required
-      await setBearerAuthToObject(localVarHeaderParameter, configuration)
-
-      if (currentPage !== undefined) {
-        localVarQueryParameter['currentPage'] = currentPage
-      }
-
-      if (pagination !== undefined) {
-        localVarQueryParameter['pagination'] = pagination
-      }
-
-      if (states) {
-        localVarQueryParameter['states'] = states
-      }
-
-      if (userCanEdit !== undefined) {
-        localVarQueryParameter['userCanEdit'] = userCanEdit
-      }
-
-      if (formDefinitionSlug !== undefined) {
-        localVarQueryParameter['formDefinitionSlug'] = formDefinitionSlug
-      }
-
-      localVarHeaderParameter['Accept'] = 'application/json'
-
-      setSearchParams(localVarUrlObj, localVarQueryParameter)
-      let headersFromBaseOptions = baseOptions && baseOptions.headers ? baseOptions.headers : {}
-      localVarRequestOptions.headers = {
-        ...localVarHeaderParameter,
-        ...headersFromBaseOptions,
-        ...options.headers,
-      }
-
-      return {
-        url: toPathString(localVarUrlObj),
-        options: localVarRequestOptions,
-      }
-    },
-    /**
-     * This endpoint is used for updating from and sending it to NASES. First is form updated then send to rabbitmq, then is controlled if everything is okay and files are scanned and after that is send to NASES
-     * @summary
-     * @param {string} formId
-     * @param {UpdateFormRequestDto} updateFormRequestDto
-     * @param {*} [options] Override http request option.
-     * @deprecated
-     * @throws {RequiredError}
-     */
-    nasesControllerSendAndUpdateForm: async (
-      formId: string,
-      updateFormRequestDto: UpdateFormRequestDto,
-      options: RawAxiosRequestConfig = {},
-    ): Promise<RequestArgs> => {
-      // verify required parameter 'formId' is not null or undefined
-      assertParamExists('nasesControllerSendAndUpdateForm', 'formId', formId)
-      // verify required parameter 'updateFormRequestDto' is not null or undefined
-      assertParamExists(
-        'nasesControllerSendAndUpdateForm',
-        'updateFormRequestDto',
-        updateFormRequestDto,
-      )
-      const localVarPath = `/nases/send-and-update-form/{formId}`.replace(
-        '{formId}',
-        encodeURIComponent(String(formId)),
-      )
-      // use dummy base URL string because the URL constructor only accepts absolute URLs.
-      const localVarUrlObj = new URL(localVarPath, DUMMY_BASE_URL)
-      let baseOptions
-      if (configuration) {
-        baseOptions = configuration.baseOptions
-      }
-
-      const localVarRequestOptions = { method: 'POST', ...baseOptions, ...options }
-      const localVarHeaderParameter = {} as any
-      const localVarQueryParameter = {} as any
-
-      // authentication cognitoGuestIdentityId required
-      await setApiKeyToObject(localVarHeaderParameter, 'X-Cognito-Guest-Identity-Id', configuration)
-
-      // authentication bearer required
-      // http bearer authentication required
-      await setBearerAuthToObject(localVarHeaderParameter, configuration)
-
-      localVarHeaderParameter['Content-Type'] = 'application/json'
-      localVarHeaderParameter['Accept'] = 'application/json'
-
-      setSearchParams(localVarUrlObj, localVarQueryParameter)
-      let headersFromBaseOptions = baseOptions && baseOptions.headers ? baseOptions.headers : {}
-      localVarRequestOptions.headers = {
-        ...localVarHeaderParameter,
-        ...headersFromBaseOptions,
-        ...options.headers,
-      }
-      localVarRequestOptions.data = serializeDataIfNeeded(
-        updateFormRequestDto,
-        localVarRequestOptions,
-        configuration,
-      )
-
-      return {
-        url: toPathString(localVarUrlObj),
-        options: localVarRequestOptions,
-      }
-    },
-    /**
-     * This endpoint is used for updating from and sending it to NASES. First is form updated then send to rabbitmq, then is controlled if everything is okay and files are scanned and after that is send to NASES
-     * @summary
-     * @param {string} formId
-     * @param {EidUpdateSendFormRequestDto} eidUpdateSendFormRequestDto
-     * @param {*} [options] Override http request option.
-     * @deprecated
-     * @throws {RequiredError}
-     */
-    nasesControllerSendAndUpdateFormEid: async (
-      formId: string,
-      eidUpdateSendFormRequestDto: EidUpdateSendFormRequestDto,
-      options: RawAxiosRequestConfig = {},
-    ): Promise<RequestArgs> => {
-      // verify required parameter 'formId' is not null or undefined
-      assertParamExists('nasesControllerSendAndUpdateFormEid', 'formId', formId)
-      // verify required parameter 'eidUpdateSendFormRequestDto' is not null or undefined
-      assertParamExists(
-        'nasesControllerSendAndUpdateFormEid',
-        'eidUpdateSendFormRequestDto',
-        eidUpdateSendFormRequestDto,
-      )
-      const localVarPath = `/nases/eid/send-and-update-form/{formId}`.replace(
-        '{formId}',
-        encodeURIComponent(String(formId)),
-      )
-      // use dummy base URL string because the URL constructor only accepts absolute URLs.
-      const localVarUrlObj = new URL(localVarPath, DUMMY_BASE_URL)
-      let baseOptions
-      if (configuration) {
-        baseOptions = configuration.baseOptions
-      }
-
-      const localVarRequestOptions = { method: 'POST', ...baseOptions, ...options }
-      const localVarHeaderParameter = {} as any
-      const localVarQueryParameter = {} as any
-
-      // authentication cognitoGuestIdentityId required
-      await setApiKeyToObject(localVarHeaderParameter, 'X-Cognito-Guest-Identity-Id', configuration)
-
-      // authentication bearer required
-      // http bearer authentication required
-      await setBearerAuthToObject(localVarHeaderParameter, configuration)
-
-      localVarHeaderParameter['Content-Type'] = 'application/json'
-      localVarHeaderParameter['Accept'] = 'application/json'
-
-      setSearchParams(localVarUrlObj, localVarQueryParameter)
-      let headersFromBaseOptions = baseOptions && baseOptions.headers ? baseOptions.headers : {}
-      localVarRequestOptions.headers = {
-        ...localVarHeaderParameter,
-        ...headersFromBaseOptions,
-        ...options.headers,
-      }
-      localVarRequestOptions.data = serializeDataIfNeeded(
-        eidUpdateSendFormRequestDto,
-        localVarRequestOptions,
-        configuration,
-      )
-
-      return {
-        url: toPathString(localVarUrlObj),
-        options: localVarRequestOptions,
-      }
-    },
-    /**
-     * Update form
-     * @summary
-     * @param {string} formId
-     * @param {UpdateFormRequestDto} updateFormRequestDto
-     * @param {*} [options] Override http request option.
-     * @deprecated
-     * @throws {RequiredError}
-     */
-    nasesControllerUpdateForm: async (
-      formId: string,
-      updateFormRequestDto: UpdateFormRequestDto,
-      options: RawAxiosRequestConfig = {},
-    ): Promise<RequestArgs> => {
-      // verify required parameter 'formId' is not null or undefined
-      assertParamExists('nasesControllerUpdateForm', 'formId', formId)
-      // verify required parameter 'updateFormRequestDto' is not null or undefined
-      assertParamExists('nasesControllerUpdateForm', 'updateFormRequestDto', updateFormRequestDto)
-      const localVarPath = `/nases/update-form/{formId}`.replace(
-        '{formId}',
-        encodeURIComponent(String(formId)),
-      )
-      // use dummy base URL string because the URL constructor only accepts absolute URLs.
-      const localVarUrlObj = new URL(localVarPath, DUMMY_BASE_URL)
-      let baseOptions
-      if (configuration) {
-        baseOptions = configuration.baseOptions
-      }
-
-      const localVarRequestOptions = { method: 'POST', ...baseOptions, ...options }
-      const localVarHeaderParameter = {} as any
-      const localVarQueryParameter = {} as any
-
-      // authentication cognitoGuestIdentityId required
-      await setApiKeyToObject(localVarHeaderParameter, 'X-Cognito-Guest-Identity-Id', configuration)
-
-      // authentication bearer required
-      // http bearer authentication required
-      await setBearerAuthToObject(localVarHeaderParameter, configuration)
-
-      localVarHeaderParameter['Content-Type'] = 'application/json'
-      localVarHeaderParameter['Accept'] = 'application/json'
-
-      setSearchParams(localVarUrlObj, localVarQueryParameter)
-      let headersFromBaseOptions = baseOptions && baseOptions.headers ? baseOptions.headers : {}
-      localVarRequestOptions.headers = {
-        ...localVarHeaderParameter,
-        ...headersFromBaseOptions,
-        ...options.headers,
-      }
-      localVarRequestOptions.data = serializeDataIfNeeded(
-        updateFormRequestDto,
-        localVarRequestOptions,
-        configuration,
-      )
-
-      return {
-        url: toPathString(localVarUrlObj),
-        options: localVarRequestOptions,
-      }
-    },
-  }
-}
-
-/**
- * NasesApi - functional programming interface
- */
-export const NasesApiFp = function (configuration?: Configuration) {
-  const localVarAxiosParamCreator = NasesApiAxiosParamCreator(configuration)
-  return {
-    /**
-     * Archive form (hide from user but keep in database)
-     * @summary
-     * @param {string} formId
-     * @param {*} [options] Override http request option.
-     * @deprecated
-     * @throws {RequiredError}
-     */
-    async nasesControllerDeleteForm(
-      formId: string,
-      options?: RawAxiosRequestConfig,
-    ): Promise<(axios?: AxiosInstance, basePath?: string) => AxiosPromise<void>> {
-      const localVarAxiosArgs = await localVarAxiosParamCreator.nasesControllerDeleteForm(
-        formId,
-        options,
-      )
-      const localVarOperationServerIndex = configuration?.serverIndex ?? 0
-      const localVarOperationServerBasePath =
-        operationServerMap['NasesApi.nasesControllerDeleteForm']?.[localVarOperationServerIndex]
-          ?.url
-      return (axios, basePath) =>
-        createRequestFunction(
-          localVarAxiosArgs,
-          globalAxios,
-          BASE_PATH,
-          configuration,
-        )(axios, localVarOperationServerBasePath || basePath)
-    },
-    /**
-     * Return form by ID and by logged user
-     * @summary
-     * @param {string} formId
-     * @param {*} [options] Override http request option.
-     * @deprecated
-     * @throws {RequiredError}
-     */
-    async nasesControllerGetForm(
-      formId: string,
-      options?: RawAxiosRequestConfig,
-    ): Promise<(axios?: AxiosInstance, basePath?: string) => AxiosPromise<GetFormResponseDto>> {
-      const localVarAxiosArgs = await localVarAxiosParamCreator.nasesControllerGetForm(
-        formId,
-        options,
-      )
-      const localVarOperationServerIndex = configuration?.serverIndex ?? 0
-      const localVarOperationServerBasePath =
-        operationServerMap['NasesApi.nasesControllerGetForm']?.[localVarOperationServerIndex]?.url
-      return (axios, basePath) =>
-        createRequestFunction(
-          localVarAxiosArgs,
-          globalAxios,
-          BASE_PATH,
-          configuration,
-        )(axios, localVarOperationServerBasePath || basePath)
-    },
-    /**
-     * Get paginated forms
-     * @summary Get paginated forms
-     * @param {string} [currentPage] Page number
-     * @param {string} [pagination] Number of items per page
-     * @param {Array<FormState>} [states] Forms in which states are searched - when omitted, all forms of the user are searched
-     * @param {boolean} [userCanEdit] Get only forms in such a state, that user can still edit it.
-     * @param {string} [formDefinitionSlug] Slug of the form definition
-     * @param {*} [options] Override http request option.
-     * @deprecated
-     * @throws {RequiredError}
-     */
-    async nasesControllerGetForms(
-      currentPage?: string,
-      pagination?: string,
-      states?: Array<FormState>,
-      userCanEdit?: boolean,
-      formDefinitionSlug?: string,
-      options?: RawAxiosRequestConfig,
-    ): Promise<(axios?: AxiosInstance, basePath?: string) => AxiosPromise<GetFormsResponseDto>> {
-      const localVarAxiosArgs = await localVarAxiosParamCreator.nasesControllerGetForms(
-        currentPage,
-        pagination,
-        states,
-        userCanEdit,
-        formDefinitionSlug,
-        options,
-      )
-      const localVarOperationServerIndex = configuration?.serverIndex ?? 0
-      const localVarOperationServerBasePath =
-        operationServerMap['NasesApi.nasesControllerGetForms']?.[localVarOperationServerIndex]?.url
-      return (axios, basePath) =>
-        createRequestFunction(
-          localVarAxiosArgs,
-          globalAxios,
-          BASE_PATH,
-          configuration,
-        )(axios, localVarOperationServerBasePath || basePath)
-    },
-    /**
-     * This endpoint is used for updating from and sending it to NASES. First is form updated then send to rabbitmq, then is controlled if everything is okay and files are scanned and after that is send to NASES
-     * @summary
-     * @param {string} formId
-     * @param {UpdateFormRequestDto} updateFormRequestDto
-     * @param {*} [options] Override http request option.
-     * @deprecated
-     * @throws {RequiredError}
-     */
-    async nasesControllerSendAndUpdateForm(
-      formId: string,
-      updateFormRequestDto: UpdateFormRequestDto,
-      options?: RawAxiosRequestConfig,
-    ): Promise<(axios?: AxiosInstance, basePath?: string) => AxiosPromise<SendFormResponseDto>> {
-      const localVarAxiosArgs = await localVarAxiosParamCreator.nasesControllerSendAndUpdateForm(
-        formId,
-        updateFormRequestDto,
-        options,
-      )
-      const localVarOperationServerIndex = configuration?.serverIndex ?? 0
-      const localVarOperationServerBasePath =
-        operationServerMap['NasesApi.nasesControllerSendAndUpdateForm']?.[
-          localVarOperationServerIndex
-        ]?.url
-      return (axios, basePath) =>
-        createRequestFunction(
-          localVarAxiosArgs,
-          globalAxios,
-          BASE_PATH,
-          configuration,
-        )(axios, localVarOperationServerBasePath || basePath)
-    },
-    /**
-     * This endpoint is used for updating from and sending it to NASES. First is form updated then send to rabbitmq, then is controlled if everything is okay and files are scanned and after that is send to NASES
-     * @summary
-     * @param {string} formId
-     * @param {EidUpdateSendFormRequestDto} eidUpdateSendFormRequestDto
-     * @param {*} [options] Override http request option.
-     * @deprecated
-     * @throws {RequiredError}
-     */
-    async nasesControllerSendAndUpdateFormEid(
-      formId: string,
-      eidUpdateSendFormRequestDto: EidUpdateSendFormRequestDto,
-      options?: RawAxiosRequestConfig,
-    ): Promise<(axios?: AxiosInstance, basePath?: string) => AxiosPromise<SendFormResponseDto>> {
-      const localVarAxiosArgs = await localVarAxiosParamCreator.nasesControllerSendAndUpdateFormEid(
-        formId,
-        eidUpdateSendFormRequestDto,
-        options,
-      )
-      const localVarOperationServerIndex = configuration?.serverIndex ?? 0
-      const localVarOperationServerBasePath =
-        operationServerMap['NasesApi.nasesControllerSendAndUpdateFormEid']?.[
-          localVarOperationServerIndex
-        ]?.url
-      return (axios, basePath) =>
-        createRequestFunction(
-          localVarAxiosArgs,
-          globalAxios,
-          BASE_PATH,
-          configuration,
-        )(axios, localVarOperationServerBasePath || basePath)
-    },
-    /**
-     * Update form
-     * @summary
-     * @param {string} formId
-     * @param {UpdateFormRequestDto} updateFormRequestDto
-     * @param {*} [options] Override http request option.
-     * @deprecated
-     * @throws {RequiredError}
-     */
-    async nasesControllerUpdateForm(
-      formId: string,
-      updateFormRequestDto: UpdateFormRequestDto,
-      options?: RawAxiosRequestConfig,
-    ): Promise<(axios?: AxiosInstance, basePath?: string) => AxiosPromise<UpdateFormResponseDto>> {
-      const localVarAxiosArgs = await localVarAxiosParamCreator.nasesControllerUpdateForm(
-        formId,
-        updateFormRequestDto,
-        options,
-      )
-      const localVarOperationServerIndex = configuration?.serverIndex ?? 0
-      const localVarOperationServerBasePath =
-        operationServerMap['NasesApi.nasesControllerUpdateForm']?.[localVarOperationServerIndex]
-          ?.url
-      return (axios, basePath) =>
-        createRequestFunction(
-          localVarAxiosArgs,
-          globalAxios,
-          BASE_PATH,
-          configuration,
-        )(axios, localVarOperationServerBasePath || basePath)
-    },
-  }
-}
-
-/**
- * NasesApi - factory interface
- */
-export const NasesApiFactory = function (
-  configuration?: Configuration,
-  basePath?: string,
-  axios?: AxiosInstance,
-) {
-  const localVarFp = NasesApiFp(configuration)
-  return {
-    /**
-     * Archive form (hide from user but keep in database)
-     * @summary
-     * @param {string} formId
-     * @param {*} [options] Override http request option.
-     * @deprecated
-     * @throws {RequiredError}
-     */
-    nasesControllerDeleteForm(formId: string, options?: RawAxiosRequestConfig): AxiosPromise<void> {
-      return localVarFp
-        .nasesControllerDeleteForm(formId, options)
-        .then((request) => request(axios, basePath))
-    },
-    /**
-     * Return form by ID and by logged user
-     * @summary
-     * @param {string} formId
-     * @param {*} [options] Override http request option.
-     * @deprecated
-     * @throws {RequiredError}
-     */
-    nasesControllerGetForm(
-      formId: string,
-      options?: RawAxiosRequestConfig,
-    ): AxiosPromise<GetFormResponseDto> {
-      return localVarFp
-        .nasesControllerGetForm(formId, options)
-        .then((request) => request(axios, basePath))
-    },
-    /**
-     * Get paginated forms
-     * @summary Get paginated forms
-     * @param {string} [currentPage] Page number
-     * @param {string} [pagination] Number of items per page
-     * @param {Array<FormState>} [states] Forms in which states are searched - when omitted, all forms of the user are searched
-     * @param {boolean} [userCanEdit] Get only forms in such a state, that user can still edit it.
-     * @param {string} [formDefinitionSlug] Slug of the form definition
-     * @param {*} [options] Override http request option.
-     * @deprecated
-     * @throws {RequiredError}
-     */
-    nasesControllerGetForms(
-      currentPage?: string,
-      pagination?: string,
-      states?: Array<FormState>,
-      userCanEdit?: boolean,
-      formDefinitionSlug?: string,
-      options?: RawAxiosRequestConfig,
-    ): AxiosPromise<GetFormsResponseDto> {
-      return localVarFp
-        .nasesControllerGetForms(
-          currentPage,
-          pagination,
-          states,
-          userCanEdit,
-          formDefinitionSlug,
-          options,
-        )
-        .then((request) => request(axios, basePath))
-    },
-    /**
-     * This endpoint is used for updating from and sending it to NASES. First is form updated then send to rabbitmq, then is controlled if everything is okay and files are scanned and after that is send to NASES
-     * @summary
-     * @param {string} formId
-     * @param {UpdateFormRequestDto} updateFormRequestDto
-     * @param {*} [options] Override http request option.
-     * @deprecated
-     * @throws {RequiredError}
-     */
-    nasesControllerSendAndUpdateForm(
-      formId: string,
-      updateFormRequestDto: UpdateFormRequestDto,
-      options?: RawAxiosRequestConfig,
-    ): AxiosPromise<SendFormResponseDto> {
-      return localVarFp
-        .nasesControllerSendAndUpdateForm(formId, updateFormRequestDto, options)
-        .then((request) => request(axios, basePath))
-    },
-    /**
-     * This endpoint is used for updating from and sending it to NASES. First is form updated then send to rabbitmq, then is controlled if everything is okay and files are scanned and after that is send to NASES
-     * @summary
-     * @param {string} formId
-     * @param {EidUpdateSendFormRequestDto} eidUpdateSendFormRequestDto
-     * @param {*} [options] Override http request option.
-     * @deprecated
-     * @throws {RequiredError}
-     */
-    nasesControllerSendAndUpdateFormEid(
-      formId: string,
-      eidUpdateSendFormRequestDto: EidUpdateSendFormRequestDto,
-      options?: RawAxiosRequestConfig,
-    ): AxiosPromise<SendFormResponseDto> {
-      return localVarFp
-        .nasesControllerSendAndUpdateFormEid(formId, eidUpdateSendFormRequestDto, options)
-        .then((request) => request(axios, basePath))
-    },
-    /**
-     * Update form
-     * @summary
-     * @param {string} formId
-     * @param {UpdateFormRequestDto} updateFormRequestDto
-     * @param {*} [options] Override http request option.
-     * @deprecated
-     * @throws {RequiredError}
-     */
-    nasesControllerUpdateForm(
-      formId: string,
-      updateFormRequestDto: UpdateFormRequestDto,
-      options?: RawAxiosRequestConfig,
-    ): AxiosPromise<UpdateFormResponseDto> {
-      return localVarFp
-        .nasesControllerUpdateForm(formId, updateFormRequestDto, options)
-        .then((request) => request(axios, basePath))
-    },
-  }
-}
-
-/**
- * NasesApi - object-oriented interface
- */
-export class NasesApi extends BaseAPI {
-  /**
-   * Archive form (hide from user but keep in database)
-   * @summary
-   * @param {string} formId
-   * @param {*} [options] Override http request option.
-   * @deprecated
-   * @throws {RequiredError}
-   */
-  public nasesControllerDeleteForm(formId: string, options?: RawAxiosRequestConfig) {
-    return NasesApiFp(this.configuration)
-      .nasesControllerDeleteForm(formId, options)
-      .then((request) => request(this.axios, this.basePath))
-  }
-
-  /**
-   * Return form by ID and by logged user
-   * @summary
-   * @param {string} formId
-   * @param {*} [options] Override http request option.
-   * @deprecated
-   * @throws {RequiredError}
-   */
-  public nasesControllerGetForm(formId: string, options?: RawAxiosRequestConfig) {
-    return NasesApiFp(this.configuration)
-      .nasesControllerGetForm(formId, options)
-      .then((request) => request(this.axios, this.basePath))
-  }
-
-  /**
-   * Get paginated forms
-   * @summary Get paginated forms
-   * @param {string} [currentPage] Page number
-   * @param {string} [pagination] Number of items per page
-   * @param {Array<FormState>} [states] Forms in which states are searched - when omitted, all forms of the user are searched
-   * @param {boolean} [userCanEdit] Get only forms in such a state, that user can still edit it.
-   * @param {string} [formDefinitionSlug] Slug of the form definition
-   * @param {*} [options] Override http request option.
-   * @deprecated
-   * @throws {RequiredError}
-   */
-  public nasesControllerGetForms(
-    currentPage?: string,
-    pagination?: string,
-    states?: Array<FormState>,
-    userCanEdit?: boolean,
-    formDefinitionSlug?: string,
-    options?: RawAxiosRequestConfig,
-  ) {
-    return NasesApiFp(this.configuration)
-      .nasesControllerGetForms(
-        currentPage,
-        pagination,
-        states,
-        userCanEdit,
-        formDefinitionSlug,
-        options,
-      )
-      .then((request) => request(this.axios, this.basePath))
-  }
-
-  /**
-   * This endpoint is used for updating from and sending it to NASES. First is form updated then send to rabbitmq, then is controlled if everything is okay and files are scanned and after that is send to NASES
-   * @summary
-   * @param {string} formId
-   * @param {UpdateFormRequestDto} updateFormRequestDto
-   * @param {*} [options] Override http request option.
-   * @deprecated
-   * @throws {RequiredError}
-   */
-  public nasesControllerSendAndUpdateForm(
-    formId: string,
-    updateFormRequestDto: UpdateFormRequestDto,
-    options?: RawAxiosRequestConfig,
-  ) {
-    return NasesApiFp(this.configuration)
-      .nasesControllerSendAndUpdateForm(formId, updateFormRequestDto, options)
-      .then((request) => request(this.axios, this.basePath))
-  }
-
-  /**
-   * This endpoint is used for updating from and sending it to NASES. First is form updated then send to rabbitmq, then is controlled if everything is okay and files are scanned and after that is send to NASES
-   * @summary
-   * @param {string} formId
-   * @param {EidUpdateSendFormRequestDto} eidUpdateSendFormRequestDto
-   * @param {*} [options] Override http request option.
-   * @deprecated
-   * @throws {RequiredError}
-   */
-  public nasesControllerSendAndUpdateFormEid(
-    formId: string,
-    eidUpdateSendFormRequestDto: EidUpdateSendFormRequestDto,
-    options?: RawAxiosRequestConfig,
-  ) {
-    return NasesApiFp(this.configuration)
-      .nasesControllerSendAndUpdateFormEid(formId, eidUpdateSendFormRequestDto, options)
-      .then((request) => request(this.axios, this.basePath))
-  }
-
-  /**
-   * Update form
-   * @summary
-   * @param {string} formId
-   * @param {UpdateFormRequestDto} updateFormRequestDto
-   * @param {*} [options] Override http request option.
-   * @deprecated
-   * @throws {RequiredError}
-   */
-  public nasesControllerUpdateForm(
-    formId: string,
-    updateFormRequestDto: UpdateFormRequestDto,
-    options?: RawAxiosRequestConfig,
-  ) {
-    return NasesApiFp(this.configuration)
-      .nasesControllerUpdateForm(formId, updateFormRequestDto, options)
-      .then((request) => request(this.axios, this.basePath))
-  }
-}
-
-/**
  * SignerApi - axios parameter creator
  */
 export const SignerApiAxiosParamCreator = function (configuration?: Configuration) {
