+++ openapi-clients/forms/client.ts
@@ -8,7 +8,6 @@ import {
   FormsV2ApiFactory,
   GinisApiFactory,
   HealthcheckApiFactory,
-  NasesApiFactory,
   SignerApiFactory,
   StatusesApiFactory,
   WebhookApiFactory,
@@ -42,7 +41,6 @@ export const createFormsClient = ({
     ...FormsV2ApiFactory(...args),
     ...GinisApiFactory(...args),
     ...HealthcheckApiFactory(...args),
-    ...NasesApiFactory(...args),
     ...SignerApiFactory(...args),
     ...StatusesApiFactory(...args),
     ...WebhookApiFactory(...args),
